Split Processor's three self-contained concerns into modules
processor.rb had grown to 866 lines holding four separate jobs. Three of them
answer a question of their own and reach for little else, so they become mixins
beside the Conditional one that was already there:

  Constraints     what a gadget requires of its caller
  TrackedMemory   the memory a candidate reads and writes
  DataProcessing  what an instruction leaves in a register

Processor keeps the emulation lifecycle and the calls a candidate may cross,
which are entangled with both the registers and the constraint list, and is 331
lines.

ZERO is no longer a private constant of Conditional: the drop rules that read
back a compare it recorded now name it, rather than resolving it by having been
lexically inside the same class.

The corpus is byte-identical at levels 0/1/2 across every fixture.
